"fix: remove duplicate leg graphic file and simplify part selection code
- Found duplicate ProstheticLegGraphic.jsx in wrong folder causing sidebar issues
- Removed extra selectedPart state - now just use description for everything
- Deleted unnecessary onPartSelect prop connections
- Leg graphic now displays correctly in sidebar
- Tested all filters - everything works properly with no breaking changes
- Verified no interference with existing filter functionality

Files changed:
- Sidebar.jsx - removed selectedPart state
- Parts.jsx - removed onPartSelect prop
- Fixed import path to correct ProstheticLegGraphic.jsx"
