+++ src/components/SideBar.jsx
@@ -25,7 +25,6 @@ const SideBar = () => {
   const [assistiveDevice, setAssistiveDevice] = useState("");
   const [extremity, setExtremity] = useState("");
   const [description, setDescription] = useState("");
-  const [selectedPart, setSelectedPart] = useState(null);
   const [pediatric, setPediatric] = useState(false);
 
   const activeToggle = () => setIsSideBarActive(!isSideBarActive);
@@ -65,7 +64,6 @@ const SideBar = () => {
     setAssistiveDevice("");
     setExtremity("");
     setDescription("");
-    setSelectedPart(null);
     setPediatric(false);
   };
 
@@ -103,7 +101,6 @@ const SideBar = () => {
           setSelectedFilters={setSelectedFilters}
         />
 
-
         {assistiveDevice && (
           <Extremity extremity={extremity} setExtremity={setExtremity} />
         )}
@@ -111,28 +108,21 @@ const SideBar = () => {
         {extremity && (
           <>
             {extremity === "Lower" && (
-              <div>
-                <Parts
-                  description={description}
-                  setDescription={setDescription}
-                  selectedPart={selectedPart}
-                  onPartSelect={setSelectedPart} //added prop to update selected part
-                />
-                       <div className="w-full relative">
-                  {/* The leg graphic container */}
-                  <div className="absolute right-0 top-0 overflow-visible">
-                    {/* Scale + translation applied here */}
-                    <div className="scale-[2] origin-top-left translate-y-[35px] translate-x-[-150px]">
-                      <ProstheticLegGraphic
-                        selectedPart={selectedPart}
-                        onPartClick={setSelectedPart}
-                      />
-                    </div>
+              <div className="w-full relative">
+                <div className="absolute right-0 top-0 overflow-visible">
+                  <div className="scale-[2] origin-top-left translate-y-[50px] translate-x-[-150px]">
+                    <ProstheticLegGraphic 
+                      selectedPart={description}
+                    />
                   </div>
                 </div>
               </div>
             )}
 
+            <Parts 
+              description={description} 
+              setDescription={setDescription}
+            />
             <Pediatric pediatric={pediatric} setPediatric={setPediatric} setSelectedFilters={setSelectedFilters} />
             <Manufacturer />
             <Size />
